+++ src/main/java/com/hai/minh/ecommerce/repository/CategoryRepository.java
@@ -8,6 +8,6 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
     List<CategoryEntity> findByNameIn(Collection<String> name);
 }
